refactor(media): extract dimension, skip, tier-list, and debrief modules from comparisons service
Extract 4 modules from service.ts (1264 → 279 LOC) into comparisons/lib/:
- dimension-exclusion.ts — excludeFromDimension(), includeInDimension()
- skip-cooloff.ts — recordSkip(), isPairOnCooloff()
- tier-list.ts — getTierListMovies(), batchRecordComparisons(), submitTierList()
- debrief.ts — getDebriefOpponent(), getPendingDebriefs(),
  dismissDebriefDimension(), recordDebriefComparison()

service.ts retains core orchestrators: recordComparison, deleteComparison,
blacklistMovie + thin recordDebriefComparison wrapper (DI pattern to avoid
circular deps). All 247 tests pass unchanged. [tb-370]
